done
-i ran into some problems while connecting gallery and carousel
collection view so in process of resolving this problem i created
gallery collection view cell and gallery table view cell which are not
really necessary (i can reuse carousel cells for cast members). For now
i will leave it the way it is, and if needed i can remove redundant
cells.